Add files via upload
did some refacotring. flipped the bars to the bottom of the screen. got rid of AlgorithmButton and replaced the concept with a ComboBox. added a spinner to adjust delay realtine. implemented key bindings, ESC to close and F11 for the best fullscreen implementation ive ever written. the start of the UI is emmerging